refactor(jsx): split hydration and reconciliation into focused modules
Decompose the large hydration and reconciliation files into smaller units
for range updates, disposal, planning, and mounted-range recovery while
keeping the public reconciliation barrel and behavior unchanged.
